+++ Tests/LinkedListTests/LinkedListQueryingTests.swift
@@ -27,7 +27,7 @@ final class LinkedListQueryingTests: XCTestCase {
     func testIndexAt() throws {
         for index in 0..<vts.count {
             let vtsItem = vts[index]
-            let sutItem = sut.nodeAt(UInt(index)).item
+            let sutItem = sut.nodeAt(index).item
             
             XCTAssertEqual(sutItem, vtsItem)
         }
